moved core1 main functions into main1, moved more definitions from main into PayloadConfig, added comments to parent headers
