+++ ascendfsw/src/main.cpp
@@ -1,3 +1,8 @@
+/**
+ * @file main.cpp
+ * @brief Main functions for Core 0, responsible for reading data from sensor
+ * peripherals
+ */
 #include <Arduino.h>
 
 // error code framework
@@ -7,7 +12,6 @@
 
 // parent classes
 #include "Sensor.h"
-#include "Storage.h"
 
 // include sensor headers here
 #include "AS7331Sensor.h"
@@ -25,13 +29,7 @@
 
 // helper function definitions
 int verifySensors();
-int verifyStorage();
-int verifySensorRecovery(); 
-int verifyStorageRecovery(); 
-
-void storeData(String data);
-void storeDataPacket(uint8_t* packet); 
-
+int verifySensorRecovery();
 String readSensorData();
 void readSensorDataPacket(uint8_t* packet);
 String decodePacket(uint8_t* packet);
@@ -63,48 +61,23 @@ Sensor* sensors[] = {&rtc_sensor,     &bme_sensor,    &ina260_sensor,
                      &bme280_sensor,  &ens160_sensor, &uv_sensor_1,
                      &uv_sensor_2,    &icm_sensor,    &gps_sensor};
 //&gps_sensor};
+
 const int sensors_len = sizeof(sensors) / sizeof(sensors[0]);
 bool sensors_verify[sensors_len];
 String header_condensed = "";
 
-// include storage headers here
-#include "FlashStorage.h"
-#include "RadioStorage.h"
-#include "SDStorage.h"
-
-// storage classes
-SDStorage sd_storage;
-RadioStorage radio_storage;
-FlashStorage flash_storage;
-
-// storage array
+// remnant
 #if FLASH_SPI1 == 0
-Storage* storages[] = {&sd_storage, &radio_storage};
-#else
-Storage* storages[] = {&sd_storage, &radio_storage, &flash_storage};
+#include "FlashStorage.h"
+FlashStorage flash_storage;
 #endif
-
-const int storages_len = sizeof(storages) / sizeof(storages[0]);
-bool storages_verify[storages_len];
 
 // global variables for main
 // loop counter
 unsigned int it = 0;
 
-// packet properties
-#define MAX_PACKET_SIZE 500
-const uint8_t SYNC_BYTES[] = {0x89, 0xAB, 0xCD, 0xEF};
-
-#define PACKET_SYSTEM_TESTING 1
-
-// multicore transfer queue
-#define QT_ENTRY_SIZE 1000
-#define QT_MAX_SIZE 10
-
 queue_t qt;
 // char qt_entry[QT_ENTRY_SIZE];
-
-#define RECOVERY_SYSTEM 1
 
 /**
  * @brief Setup for core 0
@@ -123,12 +96,12 @@ void setup() {
   // setup heartbeat pins
   pinMode(HEARTBEAT_PIN_0, OUTPUT);
 
-  // verify sensors
-  #if RECOVERY_SYSTEM
-  int verified_count = verifySensorRecovery(); 
-  #else
+// verify sensors
+#if RECOVERY_SYSTEM
+  int verified_count = verifySensorRecovery();
+#else
   int verified_count = verifySensors();
-  #endif
+#endif
 
   if (verified_count == 0) {
     log_core("All sensor communications failed");
@@ -170,7 +143,7 @@ void setup() {
 
   // send header to core1
   queue_add_blocking(&qt, header.c_str());
-#endif 
+#endif
 
   pinMode(ON_BOARD_LED_PIN, OUTPUT);
   log_core("Setup done.");
@@ -243,11 +216,11 @@ Software control #if FLASH_SPI1 if (was_dumping == false) { while
 
 /**
  * @brief Uses Device abstraction to verify sensors
- * Incompatible with initial header generation 
- * 
- * @return int Number of sensors verified 
+ * Incompatible with initial header generation
+ *
+ * @return int Number of sensors verified
  */
-int verifySensorRecovery(){
+int verifySensorRecovery() {
   int count = 0;
   for (int i = 0; i < sensors_len; i++) {
     if (sensors[i]->attemptConnection()) {
@@ -324,11 +297,11 @@ void readSensorDataPacket(uint8_t* packet) {
   sensor_id = (sensor_id << 1) | 1;
   // rest of the packet
   for (int i = 0; i < sensors_len; i++) {
-    #if RECOVERY_SYSTEM
+#if RECOVERY_SYSTEM
     if (sensors[i]->attemptConnection()) {
-    #else
+#else
     if (sensors_verify[i]) {
-    #endif
+#endif
       sensors[i]->getDataPacket(sensor_id, temp_packet);
     } else {
       sensor_id <<= 1;
@@ -425,85 +398,15 @@ String decodePacket(uint8_t* packet) {
 String readSensorData() {
   String csv_row = header_condensed + "," + String(millis()) + ",";
   for (int i = 0; i < sensors_len; i++) {
-    #if RECOVERY_SYSTEM 
+#if RECOVERY_SYSTEM
     if (sensors[i]->attemptConnection()) {
-    #else
+#else
     if (sensors_verify[i]) {
-    #endif
+#endif
       csv_row += sensors[i]->getDataCSV();
     }
   }
   return csv_row;
-}
-
-/**
- * @brief Verifies the connection with each storage device, and defines the
- * header_condensed field, uses recovery system 
- *
- * @return int The number of verified storage devices
- */
-int verifyStorageRecovery() {
-  int count = 0;
-  for (int i = 0; i < storages_len; i++) {
-    if (storages[i]->attemptConnection()) {
-      log_core(storages[i]->getStorageName() + " verified.");
-      count++;
-    }
-  }
-  return count;
-}
-
-/**
- * @brief Verifies the connection with each storage device, and defines the
- * header_condensed field
- *
- * @return int The number of verified storage devices
- */
-int verifyStorage() {
-  int count = 0;
-  for (int i = 0; i < storages_len; i++) {
-    storages_verify[i] = storages[i]->verify();
-    if (storages_verify[i]) {
-      log_core(storages[i]->getStorageName() + " verified.");
-      count++;
-    }
-  }
-  return count;
-}
-
-/**
- * @brief Sends data to each storage device, assumes storage devices take care
- * of newline/data end themselves
- *
- * @param data Data in a CSV formatted string
- */
-void storeData(String data) {
-  for (int i = 0; i < storages_len; i++) {
-    #if RECOVERY_SYSTEM
-    if (storages[i]->attemptConnection()) {
-    #else
-    if (storages_verify[i]) {
-    #endif
-      storages[i]->store(data);
-    }
-  }
-}
-
-void storeDataPacket(uint8_t* packet){
-  // pull length of packet out 
-  uint16_t packet_len; 
-  memcpy(&packet_len, packet + sizeof(SYNC_BYTES) + sizeof(uint32_t), sizeof(uint16_t)); 
-
-
-  for (int i = 0; i < storages_len; i++) {
-    #if RECOVERY_SYSTEM
-    if (storages[i]->attemptConnection()) {
-    #else
-    if (storages_verify[i]) {
-    #endif
-      storages[i]->store(packet, packet_len);
-    }
-  }
 }
 
 /**
@@ -517,64 +420,4 @@ void handleDataInterface() {
     flash_storage.dump();
     last_dump = millis();
   }
-}
-
-/** -------------------------------------------------------------------
- * CORE 1 CODE ONLY AFTER THIS, DO NOT MIX CODE FOR THE CORES
- *  -------------------------------------------------------------------
- * TODO: Implement multicore for communication here
- * setup queue for data transfer between cores
- *
- */
-
-// separate 8k stacks 
-bool core1_separate_stack = true;
-
-/**
- * @brief Setup for core 1
- *
- *
- */
-void setup1() {
-  // set up heartbeat 
-  pinMode(HEARTBEAT_PIN_1, OUTPUT);
-
-  // set storages to be tried forever, with 5 second recovery factor  
-  for(size_t i = 0; i < storages_len; i++){
-    storages[i]->recoveryConfig(-1, 5000); 
-  }
-
-  // verify storage
-  log_core("Verifying storage...");
-  int verified_count = verifyStorage();
-  if (verified_count == 0) {
-    log_core("No storages verified, output will be Serial only.");
-    ErrorDisplay::instance().addCode(Error::CRITICAL_FAIL);
-  }
-
-  delay(500);  // wait for other setup to run
-
-  // // verify storage
-  // if (verifyStorage() == 0) {
-  //   log_core("No storages verified, output will be Serial only.");
-  // }
-}
-
-int it2 = 0;
-/**
- * @brief Loop for core 1
- *
- */
-void loop1() {
-  it2++;
-  digitalWrite(HEARTBEAT_PIN_1, (it2 & 0x1));
-
-  char received_data[QT_ENTRY_SIZE];
-  queue_remove_blocking(&qt, received_data);
-
-  // store csv row
-  storeData(String(received_data));
-
-  // log_core("[CORE1]\t" + queue_get_level(&qt) + String(received_data));
-
 }
